requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.0
idna==2.8
oauthlib==3.1.0
protobuf==3.14.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pytz==2020.5
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.15.0
uritemplate==3.0.0
urllib3==1.25.3
